pyproject.toml
[tool.poetry]
name = "atomic-tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
django = "^6.0.3"
python-dateutil = "^2.9.0.post0"
redis = "^7.3.0"
django-celery-beat = "^2.9.0"
celery = "^5.6.2"
python-decouple = "^3.8"
djangorestframework = "^3.16.1"
ipython = "^9.11.0"
django-cors-headers = "^4.9.0"
djangorestframework-simplejwt = "^5.5.1"
drf-spectacular = "^0.29.0"
django-filter = "^25.2"
python-telegram-bot = "^22.5"
requests = "^2.32.5"

[tool.poetry.group.test.dependencies]
pytest = "^8.4.2"
pytest-django = "^4.11.1"
pytest-cov = "^7.0.0"
factory-boy = "^3.3.3"


[tool.poetry.group.lint.dependencies]
mypy = "^1.19.1"
isort = "^8.0.1"
black = "^26.3.0"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
/(\.git)/
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'
